+++ noir/acvm-repo/brillig/src/black_box.rs
@@ -11,6 +11,8 @@ pub enum BlackBoxOp {
     Blake2s { message: HeapVector, output: HeapArray },
     /// Calculates the Keccak256 hash of the inputs.
     Keccak256 { message: HeapVector, output: HeapArray },
+    /// Keccak Permutation function of 1600 width
+    Keccakf1600 { message: HeapVector, output: HeapArray },
     /// Verifies a ECDSA signature over the secp256k1 curve.
     EcdsaSecp256k1 {
         hashed_msg: HeapVector,
@@ -42,7 +44,13 @@ pub enum BlackBoxOp {
     /// Performs scalar multiplication over the embedded curve.
     FixedBaseScalarMul { low: RegisterIndex, high: RegisterIndex, result: HeapArray },
     /// Performs addtion over the embedded curve.
-    EmbeddedCurveAdd { input1_x: RegisterIndex, input1_y: RegisterIndex, input2_x: RegisterIndex, input2_y: RegisterIndex, result: HeapArray },
+    EmbeddedCurveAdd {
+        input1_x: RegisterIndex,
+        input1_y: RegisterIndex,
+        input2_x: RegisterIndex,
+        input2_y: RegisterIndex,
+        result: HeapArray,
+    },
     /// Performs point doubling over the embedded curve.
     EmbeddedCurveDouble { input1_x: RegisterIndex, input1_y: RegisterIndex, result: HeapArray },
 }
