+++ noir/compiler/noirc_evaluator/src/brillig/brillig_ir/debug_show.rs
@@ -350,6 +350,9 @@ impl DebugShow {
             BlackBoxOp::Keccak256 { message, output } => {
                 debug_println!(self.enable_debug_trace, "  KECCAK256 {} -> {}", message, output);
             }
+            BlackBoxOp::Keccakf1600 { message, output } => {
+                debug_println!(self.enable_debug_trace, "  KECCAKF1600 {} -> {}", message, output);
+            }
             BlackBoxOp::Blake2s { message, output } => {
                 debug_println!(self.enable_debug_trace, "  BLAKE2S {} -> {}", message, output);
             }
